Test/k8s/requirements.txt
-i https://pypi.org/simple
bcrypt==3.2.2
cffi==1.15.0
cryptography==37.0.2
fabric==2.7.0
invoke==1.7.1
jinja2==3.1.2
markupsafe==2.1.1
paramiko==2.11.0
pathlib2==2.3.7.post1
pycparser==2.21
pynacl==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
